+++ src/main/java/kitchenpos/application/mapper/OrderMapper.java
@@ -1,8 +1,8 @@
 package kitchenpos.application.mapper;
 
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.OrderRequest;
-import kitchenpos.application.dto.OrderRequest.OrderLineItemRequest;
+import kitchenpos.application.dto.request.OrderRequest;
+import kitchenpos.application.dto.request.OrderRequest.OrderLineItemRequest;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.repository.OrderTableRepository;
