+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -10,12 +10,12 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import kitchenpos.SpringBootTestWithProfiles;
-import kitchenpos.application.dto.MenuProductRequest;
-import kitchenpos.application.dto.MenuRequest;
-import kitchenpos.application.dto.OrderRequest;
-import kitchenpos.application.dto.OrderRequest.OrderLineItemRequest;
-import kitchenpos.application.dto.OrderStatusRequest;
-import kitchenpos.application.dto.TableRequest;
+import kitchenpos.application.dto.request.MenuProductRequest;
+import kitchenpos.application.dto.request.MenuRequest;
+import kitchenpos.application.dto.request.OrderRequest;
+import kitchenpos.application.dto.request.OrderRequest.OrderLineItemRequest;
+import kitchenpos.application.dto.request.OrderStatusRequest;
+import kitchenpos.application.dto.request.TableRequest;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.Order;
