+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -10,8 +10,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import kitchenpos.SpringBootTestWithProfiles;
-import kitchenpos.application.dto.TableGroupRequest;
-import kitchenpos.application.dto.TableGroupRequest.OrderTableId;
+import kitchenpos.application.dto.request.TableGroupRequest;
+import kitchenpos.application.dto.request.TableGroupRequest.OrderTableId;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
