+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -8,9 +8,9 @@ import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Collections;
 import kitchenpos.SpringBootTestWithProfiles;
-import kitchenpos.application.dto.TableGroupRequest;
-import kitchenpos.application.dto.TableGroupRequest.OrderTableId;
-import kitchenpos.application.dto.TableRequest;
+import kitchenpos.application.dto.request.TableGroupRequest;
+import kitchenpos.application.dto.request.TableGroupRequest.OrderTableId;
+import kitchenpos.application.dto.request.TableRequest;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
